requirements.txt
streamlit
pdfplumber
pdf2image
pillow
pytesseract
requests
google-generativeai
python-dotenv
